pyproject.toml
[project]
name = "resume-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.5.2",
    "fastapi[standard]>=0.115.12",
    "google-auth>=2.40.1",
    "itsdangerous>=2.2.0",
    "jinja2==3.1.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
    "weasyprint==59.0",
]
